ai_server/requirements.txt
fastapi==0.121.1
uvicorn[standard]==0.38.0
torch==2.9.0
torchvision==0.24.0
ultralytics==8.3.227
pillow==12.0.0
python-multipart==0.0.20
pydantic-settings==2.12.0

